refactor(ui): extract feed/sanitize.ts — HTML/Markdown sanitizer
